requirements.txt
pandas>=2.0.0
chardet>=5.0.0
openpyxl>=3.1.0
pyarrow>=12.0.0
pydantic>=2.0.0
Pillow>=9.0.0
black>=23.0.0
ruff>=0.1.0
pytest>=7.0.0
